Refactor feedback widget instantiation
