+++ src/components/Widgets/FeedbackWidget/components/CloseButton.js
