+++ src/components/Widgets/FeedbackWidget/components/StarRating.js
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from '@emotion/styled';
 import { useFeedbackState } from '../context';
 import { css } from '@emotion/core';
-import { isBrowser } from '../../../utils/is-browser';
+import { isBrowser } from '../../../../utils/is-browser';
 import { uiColors } from '@leafygreen-ui/palette';
 import loadable from '@loadable/component';
