+++ src/components/Widgets/FeedbackWidget/icons.js
