+++ src/components/Widgets/FeedbackWidget/views/QualifiersView.js
