+++ src/components/Widgets/FeedbackWidget/views/RatingView.js
