+++ src/components/layout.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import SiteMetadata from './site-metadata';
 import { TabContext } from './tab-context';
+import Widgets from './Widgets';
 import { findAllKeyValuePairs } from '../utils/find-all-key-value-pairs';
 import { getNestedValue } from '../utils/get-nested-value';
 import { getPlaintext } from '../utils/get-plaintext';
@@ -131,7 +132,7 @@ export default class DefaultLayout extends Component {
   render() {
     const {
       children,
-      pageContext: { metadata, slug },
+      pageContext: { location, metadata, slug },
     } = this.props;
     const { pillstrips } = this.state;
     const lookup = slug === '/' ? 'index' : slug;
@@ -139,12 +140,14 @@ export default class DefaultLayout extends Component {
     const pageTitle = getPlaintext(getNestedValue(['slugToTitle', lookup], metadata));
     return (
       <TabContext.Provider value={{ ...this.state, setActiveTab: this.setActiveTab }}>
-        <SiteMetadata siteTitle={siteTitle} pageTitle={pageTitle} />
-        {React.cloneElement(children, {
-          pillstrips,
-          addPillstrip: this.addPillstrip,
-          footnotes: this.footnotes,
-        })}
+        <Widgets location={location} pageTitle={pageTitle} publishedBranches={metadata.publishedBranches} slug={slug}>
+          <SiteMetadata siteTitle={siteTitle} pageTitle={pageTitle} />
+          {React.cloneElement(children, {
+            pillstrips,
+            addPillstrip: this.addPillstrip,
+            footnotes: this.footnotes,
+          })}
+        </Widgets>
       </TabContext.Provider>
     );
   }
