+++ src/hooks/use-site-metadata.js
@@ -6,7 +6,11 @@ export const useSiteMetadata = () => {
       query SiteMetaData {
         site {
           siteMetadata {
+            commitHash
+            database
             parserBranch
+            parserUser
+            patchId
             project
             snootyBranch
             user
