+++ src/templates/document.js
@@ -10,9 +10,6 @@ import DocumentBody from '../components/DocumentBody';
 import { useWindowSize } from '../hooks/use-window-size.js';
 import style from '../styles/navigation.module.css';
 import { isBrowser } from '../utils/is-browser.js';
-import { getPlaintext } from '../utils/get-plaintext.js';
-
-import { FeedbackProvider, FeedbackForm, FeedbackTab, useFeedbackData } from '../components/FeedbackWidget';
 
 const Document = props => {
   const {
@@ -24,14 +21,6 @@ const Document = props => {
     location,
     ...rest
   } = props;
-
-  const title = getPlaintext(getNestedValue([slug], slugTitleMapping));
-  const feedbackData = useFeedbackData({
-    slug,
-    title: title || 'Home',
-    url: location.href,
-    publishedBranches,
-  });
 
   const windowSize = useWindowSize();
   const minWindowWidth = 1093; /* Specific value from docs-tools/themes/mongodb/src/css/mongodb-base.css */
@@ -45,9 +34,7 @@ const Document = props => {
   };
 
   return (
-    <FeedbackProvider page={feedbackData}>
-      <FeedbackTab />
-      <FeedbackForm />
+    <React.Fragment>
       <Navbar />
       <div className="content">
         <div>
@@ -82,7 +69,7 @@ const Document = props => {
           </div>
         </div>
       </div>
-    </FeedbackProvider>
+    </React.Fragment>
   );
 };
