+++ src/templates/guide.js
@@ -6,7 +6,6 @@ import Footer from '../components/Footer';
 import GuideBreadcrumbs from '../components/GuideBreadcrumbs';
 import GuideSection from '../components/GuideSection';
 import GuideHeading from '../components/GuideHeading';
-import Widgets from '../components/Widgets/Widgets';
 import { LANGUAGES, DEPLOYMENTS, SECTION_NAME_MAPPING } from '../constants';
 import { findKeyValuePair } from '../utils/find-key-value-pair';
 import { throttle } from '../utils/throttle';
@@ -172,7 +171,6 @@ export default class Guide extends Component {
               <Footer />
             </div>
           </div>
-          {!process.env.PREVIEW_PAGE && <Widgets guideName={pageSlug} snootyStitchId={pageContext.snootyStitchId} />}
         </div>
       </React.Fragment>
     );
@@ -187,7 +185,6 @@ Guide.propTypes = {
         children: PropTypes.array,
       }).isRequired,
     }).isRequired,
-    snootyStitchId: PropTypes.string.isRequired,
   }).isRequired,
   path: PropTypes.string.isRequired,
   pillstrips: PropTypes.objectOf(PropTypes.object),
